requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
psycopg2-binary
telethon

# Use CPU-only torch and torchvision
torch==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

ultralytics
dbt-core
dbt-postgres
dagster
dagit
pandas
numpy
sqlalchemy
pydantic
loguru
